requirements.txt
asgiref==3.5.2
certifi==2022.6.15       
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.15
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
Pillow==9.2.0
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1